pyproject.toml
[tool.poetry]
name = "dogs"
version = "0.1.0"
description = ""
authors = ["daniel <ntdu198@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
pydantic-settings = "^2.4.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
django-debug-toolbar = "^4.4.6"
drf-yasg = "^1.21.7"
setuptools = "^74.1.2"
django-storages = "^1.14.4"
boto3 = "^1.35.14"
django-filter = "^24.3"
celery = "^5.4.0"
redis = "^5.0.8"
httpx = "^0.27.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
black = "^24.8.0"
flake8 = "^7.1.1"
coverage = "^7.6.1"
pytest = "^8.3.2"
respx = "^0.21.1"
moto = {extras = ["all"], version = "^5.0.13"}

[tool.black]
line-length = 120

[tool.flake8]
max-line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
